Refactor sign-up functionality and user service
