+++ internal/business/user/user_service.go
@@ -3,13 +3,13 @@ package user
 import (
 	"context"
 
-	"github.com/bajra-manandhar17/personal-finance-app/internal/model"
-	"github.com/bajra-manandhar17/personal-finance-app/internal/repository/userrepo"
+	"github.com/bajra-manandhar17/personal-finance-app/internal/db/model"
+	"github.com/bajra-manandhar17/personal-finance-app/internal/db/query"
 )
 
 type UserService interface {
 	RegisterNewUser(ctx context.Context, req RegisterNewUserReq) error
-	GetUser(ctx context.Context, userId string) (*model.User, error)
+	GetUser(ctx context.Context, userId string) (*model.Users, error)
 }
 
 type RegisterNewUserReq struct {
@@ -17,16 +17,16 @@ type RegisterNewUserReq struct {
 	Password string `json:"password" validate:"required,min=8"`
 }
 
-type UserServiceOpts struct {
-	UserRepo *userrepo.UserRepo
-}
+// type UserServiceOpts struct {
+// 	UserRepo *userrepo.UserRepo
+// }
 
-func NewUserService(opts *UserServiceOpts) UserService {
-	if opts == nil {
+func NewUserService(query *query.Query) UserService {
+	if query == nil {
 		return &UserServiceImpl{}
 	}
 
 	return &UserServiceImpl{
-		userRepo: opts.UserRepo,
+		query: query,
 	}
 }
